Open all nested menus in Docusaurus before scraping
